Added option to omit diffing lines
By providing the -olt option on startup the algorithm will omit the diffinglines
when saving the commits. This is due to a number of repositories has had very
a lot of commits which made this needed. The lines are not needed when running
with the simplebugfinder but when running with the distancefinder.